src/api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "fastapi for testing celery worker"
authors = ["Newmi <newmi1988@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.61.2"
uvicorn = "^0.12.3"
requests = "^2.25.0"
redis = "^3.5.3"
celeryworker = {path = "../celeryworker"}

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
isort = "^5.6.4"
pytest = "^6.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
